@@ -10,64 +10,49 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@invertase/react-native-apple-authentication": "^2.1.2",
-    "@react-native-async-storage/async-storage": "^1.15.4",
-    "@react-native-clipboard/clipboard": "^1.8.2",
-    "@react-native-community/datetimepicker": "^3.5.2",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-native-community/netinfo": "^6.0.0",
-    "@react-native-community/slider": "^3.0.3",
-    "@react-native-google-signin/google-signin": "^6.0.0",
-    "@react-navigation/native": "^6.0.6",
-    "@react-navigation/native-stack": "^6.2.5",
-    "@react-navigation/stack": "^6.0.11",
-    "@reduxjs/toolkit": "^1.5.1",
-    "@twotalltotems/react-native-otp-input": "^1.3.11",
+    "@react-native-async-storage/async-storage": "^1.18.1",
+    "@react-native-community/netinfo": "^9.3.10",
+    "@react-native-firebase/analytics": "^13.1.0",
+    "@react-native-firebase/app": "^16.4.1",
+    "@react-native-firebase/auth": "^16.4.3",
+    "@react-native-firebase/crashlytics": "^13.1.0",
+    "@react-native-firebase/firestore": "^15.7.0",
+    "@react-native-firebase/storage": "^16.4.6",
+    "@react-native-google-signin/google-signin": "^6.0.1",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
+    "@react-navigation/stack": "^6.3.17",
+    "@reduxjs/toolkit": "^1.9.5",
     "axios": "^0.21.1",
-    "d3-shape": "^3.0.1",
-    "moment": "^2.29.1",
-    "react": "17.0.2",
-    "react-native": "0.67.3",
-    "react-native-device-info": "^8.1.3",
-    "react-native-facebook-login": "^1.6.2",
-    "react-native-floating-label-input": "^1.3.14",
-    "react-native-fs": "^2.18.0",
-    "react-native-gesture-handler": "^1.10.3",
-    "react-native-image-blur-loading": "^1.1.5",
-    "react-native-keyboard-aware-scrollview": "^2.1.0",
-    "react-native-keyboard-aware-view": "^0.0.14",
-    "react-native-keyboard-manager": "^6.5.4-1",
-    "react-native-linear-gradient": "^2.5.6",
-    "react-native-localize": "^2.1.6",
-    "react-native-reanimated": "^2.0.0-rc.3",
-    "react-native-safe-area": "^0.5.1",
-    "react-native-safe-area-context": "^3.2.0",
-    "react-native-screens": "^3.10.2",
-    "react-native-splash-screen": "^3.2.0",
-    "react-native-svg": "^12.1.1",
-    "react-navigation": "^4.4.4",
-    "react-redux": "^7.2.4",
-    "redux": "^4.1.0",
-    "redux-saga": "^1.1.3"
+    "moment": "^2.29.4",
+    "patch-package": "^7.0.0",
+    "postinstall-postinstall": "^2.1.0",
+    "react": "18.1.0",
+    "react-native": "0.70.10",
+    "react-native-create-thumbnail": "2.0.0-rc.2",
+    "react-native-fbsdk": "^3.0.0",
+    "react-native-safe-area-context": "^4.5.2",
+    "react-native-screens": "^3.20.0",
+    "react-native-splash-screen": "^3.3.0",
+    "react-redux": "^8.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@tsconfig/react-native": "^2.0.2",
     "@types/jest": "^26.0.23",
-    "@types/react-native": "^0.66.15",
-    "@types/react-test-renderer": "^17.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "@typescript-eslint/parser": "^5.7.0",
+    "@types/react": "^18.0.21",
+    "@types/react-native": "^0.70.6",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.37.0",
+    "@typescript-eslint/parser": "^5.37.0",
     "babel-jest": "^26.6.3",
-    "eslint": "^7.14.0",
+    "eslint": "^7.32.0",
     "jest": "^26.6.3",
-    "metro-react-native-babel-preset": "^0.66.2",
-    "react-test-renderer": "17.0.2",
-    "typescript": "^4.4.4"
-  },
-  "resolutions": {
-    "@types/react": "^17"
+    "metro-react-native-babel-preset": "0.72.3",
+    "react-test-renderer": "18.1.0",
+    "typescript": "^4.8.3"
   },
   "jest": {
     "preset": "react-native",
