@@ -1,15 +1,9 @@
 /**
  * @format
  */
- import React from "react"
- import {AppRegistry} from 'react-native';
- import { Provider } from "react-redux";
- import App from './App';
- import {name as appName} from './app.json';
- import configureStore from './src/Redux/store/AppStore';
- const AppRedux = () => (
-     <Provider store={configureStore}>
-       <App />
-     </Provider>
-   );
- AppRegistry.registerComponent(appName, () => AppRedux);
+
+import {AppRegistry} from 'react-native';
+import App from './App';
+import {name as appName} from './app.json';
+
+AppRegistry.registerComponent(appName, () => App);
