@@ -1,5 +1,6 @@
 import {StackScreenProps} from '@react-navigation/stack';
 import {Dimensions, Platform, PixelRatio} from 'react-native';
+import {SocialTypeStrings} from './AppEnums';
 export const platformVersion = Platform.Version;
 export type ScreenProps = StackScreenProps<any, any>;
 export const ScreenSize = Dimensions.get('screen');
@@ -12,481 +13,345 @@ export const normalized = (value: number) =>
 export const hv = (value: number) =>
   PixelRatio.roundToNearestPixel(value * heightRatio);
 export const horizontalScreenWithMargin = ScreenSize.width - normalized(48);
-export const AppImages = {
-  home: {
-    logo: require('../Ui/assets/images/innerLogo.png'),
-    heartIcon: require('../Ui/assets/images/fav.png'),
-    cartIcon: require('../Ui/assets/images/cart.png'),
-    dropdownIcon: require('../Ui/assets/images/dropdown-Icon.png'),
-    searchIcon: require('../Ui/assets/images/search.png'),
-    filterIcon: require('../Ui/assets/images/filter.png'),
-    card1: require('../Ui/assets/images/banner1.png'),
-    card2: require('../Ui/assets/images/banner2.png'),
-    res1: require('../Ui/assets/images/resturant1.png'),
-    res2: require('../Ui/assets/images/resturant2.png'),
-    res3: require('../Ui/assets/images/resturant3.png'),
-    meal: require('../Ui/assets/images/meal.png'),
-    star: require('../Ui/assets/images/star.png'),
-    resMark: require('../Ui/assets/images/resMark.png'),
-    clock: require('../Ui/assets/images/clock.png'),
-    simpleBanner: require('../Ui/assets/images/banner3.png'),
-    back: require('../Ui/assets/images/back.png'),
-    deleteFav: require('../Ui/assets/images/deleteFav.png'),
-    noOrder : require('../Ui/assets/images/noOrder.png'),
-    locIcon : require('../Ui/assets/images/locIcon.png'),
-    startPoint : require('../Ui/assets/images/startPoint.png'),
-    acceptStatus : require('../Ui/assets/images/acceptStatus.png'),
-    progressStatus : require('../Ui/assets/images/progressStatus.png')
-  },
-  RestaurantList: {
-    bikeIcon: require('../Ui/assets/images/bikeIcon.png'),
-  },
-  Container: {
-    bottomBar: {
-      home: require('../Ui/assets/images/home.png'),
-      selectedHome: require('../Ui/assets/images/selectedHome.png'),
-      earning : require('../Ui/assets/images/earning.png'),
-      selectedEarning : require('../Ui/assets/images/selectedEarning.png'),
-      grocery: require('../Ui/assets/images/grocery.png'),
-      selectedGrocery: require('../Ui/assets/images/selectedGrocery.png'),
-      orders: require('../Ui/assets/images/orders.png'),
-      selectedOrders: require('../Ui/assets/images/selectedOrders.png'),
-      offers: require('../Ui/assets/images/offers.png'),
-      selectedOffers: require('../Ui/assets/images/selectedOffers.png'),
-      more: require('../Ui/assets/images/more.png'),
-      selectedMore: require('../Ui/assets/images/selectedMore.png'),
-      profile: require('../Ui/assets/images/profileTab.png'),
-      selectedProfile: require('../Ui/assets/images/selectedProfile.png'),
-    },
-  },
-  details: {
-    dummyDetailBg: require('../Ui/assets/images/details/dummyDetailBg.png'),
-    dummyLogo: require('../Ui/assets/images/details/dummyLogo.png'),
-    clock: require('../Ui/assets/images/details/clock.png'),
-    cock: require('../Ui/assets/images/details/cock.png'),
-    ratingStar: require('../Ui/assets/images/details/ratingStar.png'),
-    dummyLike: require('../Ui/assets/images/details/dummyLike.png'),
-    bike: require('../Ui/assets/images/details/bike.png'),
-    menu: require('../Ui/assets/images/details/menu.png'),
-    menuSelected: require('../Ui/assets/images/details/menuSelected.png'),
-    reviews: require('../Ui/assets/images/details/reviews.png'),
-    reviewsSelected: require('../Ui/assets/images/details/reviewsSelected.png'),
-    about: require('../Ui/assets/images/details/about.png'),
-    aboutSelected: require('../Ui/assets/images/details/aboutSelected.png'),
-    dummyItem: require('../Ui/assets/images/details/dummyItem.png'),
-    bannerDummy: require('../Ui/assets/images/details/bannerDummy.png'),
-    checkbox: require('../Ui/assets/images/details/checkbox.png'),
-    checkboxSelected: require('../Ui/assets/images/details/checkboxSelected.png'),
-    checkBox2Selected: require('../Ui/assets/images/details/checkBox2Selected.png'),
-    checkbox2: require('../Ui/assets/images/details/checkbox2.png'),
-    ratingStar2: require('../Ui/assets/images/details/ratingStar2.png'),
-    phoneCall: require('../Ui/assets/images/details/phoneCall.png'),
-  },
-  order: {
-    back2: require('../Ui/assets/images/back2.png'),
-    resLogo: require('../Ui/assets/images/resLogo.png'),
-    resLogo2: require('../Ui/assets/images/resLogo2.png'),
-    cancel: require('../Ui/assets/images/cancel.png'),
-    riderDummy: require('../Ui/assets/images/riderDummy.png'),
-    orderBanner: require('../Ui/assets/images/orderBanner.png'),
-  },
-  login: {
-    appIcon: require('../Ui/assets/images/miniLogo.png'),
-    appLogo: require('../Ui/assets/images/loginLogo.png'),
-    closeEye: require('../Ui/assets/images/hidePass.png'),
-    openEye: require('../Ui/assets/images/showPass.png'),
-    halfLine: require('../Ui/assets/images/halfLine.png'),
-    appleIcon: require('../Ui/assets/images/appleIcon.png'),
-    googleIcon: require('../Ui/assets/images/googleIcon.png'),
-    facebookIcon: require('../Ui/assets/images/facebookIcon.png'),
-  },
-  signUp: {
-    checked: require('../Ui/assets/images/checked.png'),
-    unchecked: require('../Ui/assets/images/unchecked.png'),
-  },
-  otp: {
-    backBtn: require('../Ui/assets/images/backBtnIcon.png'),
-  },
-  address: {
-    placesLogo: require('../Ui/assets/images/placesLogo.png'),
-    searchBtn: require('../Ui/assets/images/searchBtn.png'),
-    greenLoc: require('../Ui/assets/images/address.png'),
-    editAdd: require('../Ui/assets/images/editAdd.png'),
-    delete: require('../Ui/assets/images/delete.png'),
-    mapPin: require('../Ui/assets/images/mapPin.png'),
-  },
-  favourite: {
-    product1: require('../Ui/assets/images/product1.png'),
-    product2: require('../Ui/assets/images/product2.png'),
-    product3: require('../Ui/assets/images/product3.png'),
-    cartIcon: require('../Ui/assets/images/GoCartIcon.png'),
-  },
-  products: {
-    favIcon: require('../Ui/assets/images/fav.png'),
-    unFavIcon: require('../Ui/assets/images/unFav.png'),
-  },
-};
-export const AppFonts = {
-  InterSemiBold: 'Inter-SemiBold',
-  InterExtraBold: 'Inter-ExtraBold',
-  InterBold: 'Inter-Bold',
-  InterRegular: 'Inter',
-  InterMedium: 'Inter-Medium',
-  InterLight: 'Inter-Light',
-  SwitzerThin: 'Switzer-Thin',
-};
+export const formFieldsHeight =
+  Platform.OS == 'android' ? normalized(50) : normalized(55);
+export const isSmallDevice = ScreenSize.height < 700 ? true : false;
+export const bottomBarHeightConstant = hv(90);
+export const maxDescriptionLength = 60;
+export const maxImageSizeInBytes = 10 * 1024 * 1024; // 10MB
+
+export const AppImages = {};
+export const AppFonts = {};
+
 export const AppColors = {
-  blue: {
-    dark: '#004B87',
-    lightBlue: '#C1E4FC',
-  },
-
-  grey: {
-    light: '#C4C4C4',
-    lightOp: '#FAFAFA',
-    dark: '#6D6E72',
-    dark2: '#787878',
-  },
-
-  shadowColor: {
-    lightBlack: 'rgba(0,0,0,0.55)',
-    darkBlack: 'rgba(0,0,0,1)',
-    lightBlack2: 'rgba(0,0,0,0.3)',
-  },
-
-  yellow: {
-    dark: '#FFCD00',
-    lightYellow : "rgba(255,205,0,0.10)"
-  },
-
-  white: {
-    whiteOP: 'rgba(255,255,255,0.65)',
-    white: '#ffffff',
-    lightWhite: '#FFFFF1',
-    creamy: '#FFF9D5',
-  },
   black: {
     black: '#000000',
-    lightBlack: 'rgba(0,0,0,0.32)',
-    withShadowLight:
-      Platform.OS == 'ios' ? 'rgba(0,0,0,0.16)' : 'rgba(0,0,0,0.26)',
-    textColor: '#414141',
-    textDark: '#333333',
-    lightOP: 'rgba(0,0,0,0.75)',
-  },
-  green: {
-    dark: '#00870D',
-  },
-  red: {
-    redOp: 'rgba(243,17,17,0.10)',
-    red: 'rgba(243,17,17,1)',
-    darkRed: '#F31111',
+    darkDeep: '#090b0e',
   },
 };
 
-export const Sliders = [
+export const socialList = [
   {
     id: '1',
-    img: require('../Ui/assets/images/img1.png'),
-    title: 'part1',
-    subtitle: 'part1Desc',
+    name: SocialTypeStrings.facebook,
   },
+
   {
     id: '2',
-    img: require('../Ui/assets/images/img2.png'),
-    title: 'part2',
-    subtitle: 'part2Desc',
-  },
-  {
-    id: '3',
-    img: require('../Ui/assets/images/img3.png'),
-    title: 'part3',
-    subtitle: 'part3Desc',
-  },
-];
-export const ListTab = [
-  {
-    id: '1',
-    category: 'top',
-  },
-  {
-    id: '2',
-    category: 'recent',
-  },
-  {
-    id: '3',
-    category: 'fav',
-  },
-  {
-    id: '4',
-    category: 'new',
-  },
-  {
-    id: '5',
-    category: 'nearBy',
+    name: SocialTypeStrings.google,
   },
 ];
 
-export const bottomBarList = [
+// export const bottomBarList = [
+//   {
+//     image: AppImages.Container.BottomBar.Home,
+//     title: 'home',
+//   },
+//   {
+//     image: AppImages.Container.BottomBar.Contact,
+//     title: 'connections',
+//   },
+//   {
+//     image: '',
+//     title: 'function',
+//   },
+//   {
+//     image: AppImages.Container.BottomBar.Shop,
+//     title: 'shop',
+//   },
+//   {
+//     image: AppImages.Container.BottomBar.Scan,
+//     title: 'scan',
+//   },
+// ];
+
+export const cardColorsList = [
   {
-    selectedImage: AppImages.Container.bottomBar.selectedHome,
-    image: AppImages.Container.bottomBar.home,
-    title: 'home',
+    id: 1,
+    color: '#22A081',
+    borderColor: '#abeddd',
   },
   {
-    selectedImage: AppImages.Container.bottomBar.selectedOrders,
-    image: AppImages.Container.bottomBar.orders,
-    title: 'orders',
+    id: 2,
+    color: '#E35252',
+    borderColor: '#c52020',
   },
   {
-    selectedImage: AppImages.Container.bottomBar.selectedEarning,
-    image: AppImages.Container.bottomBar.earning,
-    title: 'earning',
+    id: 3,
+    color: '#E4A56B',
+    borderColor: '#db8a3d',
   },
   {
-    selectedImage: AppImages.Container.bottomBar.selectedProfile,
-    image: AppImages.Container.bottomBar.profile,
-    title: 'profile',
+    id: 4,
+    color: '#7000B5',
+    borderColor: '#d899ff',
   },
   {
-    selectedImage: AppImages.Container.bottomBar.selectedMore,
-    image: AppImages.Container.bottomBar.more,
-    title: 'more',
+    id: 5,
+    color: '#EEB8C9',
+    borderColor: '#e184a1',
+  },
+  {
+    id: 6,
+    color: '#DFF87C',
+    borderColor: '#b1da0b',
+  },
+  {
+    id: 7,
+    color: '#34D844',
+    borderColor: '#21ab2f',
+  },
+  {
+    id: 8,
+    color: '#3B58BD',
+    borderColor: '#9eace0',
+  },
+  {
+    id: 9,
+    color: '#FF5099',
+    borderColor: '#ff1a79',
   },
 ];
 
-///Dummy Data
-
-export const ResturantListData = [
-  {
-    id: '1',
-    pic: AppImages.home.res1,
-    title: 'Galaxy Height',
-    time: '30',
-    price: '30.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.0',
-    totalRating: '900',
-    status: 'Top',
-    timing: 'closed',
-    openingTime: '4',
-  },
-  {
-    id: '2',
-    pic: AppImages.home.res2,
-    title: 'Yummy Pizza',
-    time: '15',
-    price: '50.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '3.7',
-    totalRating: '226',
-    status: 'Top',
-  },
-  {
-    id: '3',
-    pic: AppImages.home.res3,
-    title: 'X Cafe',
-    time: '30',
-    price: '25.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.6',
-    totalRating: '309',
-    status: 'Recent',
-  },
-  {
-    id: '4',
-    pic: AppImages.home.res1,
-    title: 'Galaxy Height',
-    time: '30',
-    price: '30.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.0',
-    totalRating: '900',
-    status: ' New',
-  },
-  {
-    id: '5',
-    pic: AppImages.home.res2,
-    title: 'Yummy Pizza',
-    time: '15',
-    price: '50.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '3.7',
-    totalRating: '226',
-    status: 'Near By',
-  },
-  {
-    id: '6',
-    pic: AppImages.home.res3,
-    title: 'X Cafe',
-    time: '30',
-    price: '25.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.6',
-    totalRating: '309',
-    status: 'New',
-  },
-  {
-    id: '7',
-    pic: AppImages.home.res1,
-    title: 'Galaxy Height',
-    time: '30',
-    price: '30.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.0',
-    totalRating: '900',
-    status: 'New',
-    timing: 'closed',
-    openingTime: '4',
-  },
-  {
-    id: '8',
-    pic: AppImages.home.res2,
-    title: 'Yummy Pizza',
-    time: '15',
-    price: '50.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '3.7',
-    totalRating: '226',
-    status: 'New',
-  },
-  {
-    id: '9',
-    pic: AppImages.home.res2,
-    title: 'Yummy Pizza',
-    time: '15',
-    price: '50.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '3.7',
-    totalRating: '226',
-    status: 'New',
-  },
-  {
-    id: '10',
-    pic: AppImages.home.res3,
-    title: 'X Cafe',
-    time: '30',
-    price: '25.00',
-    resMark: AppImages.home.resMark,
-    averageRating: '4.6',
-    totalRating: '309',
-    status: 'Top',
-  },
-];
-
-export const list = [
-  {
-    id: '1',
-    pic: require('../Ui/assets/images/coffee-cup.png'),
-    title: 'coffee',
-  },
-  {
-    id: '2',
-    pic: require('../Ui/assets/images/rice.png'),
-    title: 'Rice',
-  },
-  {
-    id: '3',
-    pic: require('../Ui/assets/images/pasta.png'),
-    title: 'Pasta',
-  },
-  {
-    id: '4',
-    pic: require('../Ui/assets/images/french-fries.png'),
-    title: 'Fries',
-  },
-  {
-    id: '5',
-    pic: require('../Ui/assets/images/fried-fish.png'),
-    title: 'Fish',
-  },
-  {
-    id: '6',
-    pic: require('../Ui/assets/images/coffee-cup.png'),
-    title: 'coffee',
-  },
-  {
-    id: '7',
-    pic: require('../Ui/assets/images/rice.png'),
-    title: 'Rice',
-  },
-  {
-    id: '8',
-    pic: require('../Ui/assets/images/pasta.png'),
-    title: 'Pasta',
-  },
-  {
-    id: '9',
-    pic: require('../Ui/assets/images/french-fries.png'),
-    title: 'Fries',
-  },
-  {
-    id: '10',
-    pic: require('../Ui/assets/images/fried-fish.png'),
-    title: 'Fish',
-  },
-];
-export const moreList = [
-  {
-    name: 'profile',
-    image: require('../Ui/assets/images/profile.png'),
-  },
-  {
-    name: 'addresses',
-    image: require('../Ui/assets/images/address.png'),
-  },
-  {
-    name: 'language',
-    image: require('../Ui/assets/images/language.png'),
-  },
-  {
-    name: 'aboutUs',
-    image: require('../Ui/assets/images/about.png'),
-  },
-  {
-    name: 'contactUs',
-    image: require('../Ui/assets/images/contact.png'),
-  },
-  {
-    name: 'termsCond',
-    image: require('../Ui/assets/images/terms.png'),
-  },
-];
-//Category List Data
-
-//filtered Categories List
-
-export const filteredCategoryList = [
-  {
-    id: '1',
-    title: 'Dairy',
-  },
-  {
-    id: '2',
-    title: 'Fruits',
-  },
-  {
-    id: '3',
-    title: 'Vegetables',
-  },
-  {
-    id: '4',
-    title: 'Break Fast',
-  },
-  {
-    id: '5',
-    title: 'Baby Care',
-  },
-  {
-    id: '6',
-    title: 'Cosmetics',
-  },
-  {
-    id: '7',
-    title: 'Cleaning',
-  },
-  {
-    id: '8',
-    title: 'Frozen',
-  },
-];
+export const prefixCodes = {
+  AF: '93',
+  AE: '971',
+  AL: '355',
+  AN: '599',
+  AS: '1684',
+  AD: '376',
+  AO: '244',
+  AI: '1264',
+  AG: '1268',
+  AR: '54',
+  AM: '374',
+  AW: '297',
+  AU: '61',
+  AT: '43',
+  AZ: '994',
+  BS: '1242',
+  BH: '973',
+  BF: '226',
+  BI: '257',
+  BD: '880',
+  BB: '1246',
+  BY: '375',
+  BE: '32',
+  BZ: '501',
+  BJ: '229',
+  BM: '1441',
+  BT: '975',
+  BA: '387',
+  BW: '267',
+  BR: '55',
+  BG: '359',
+  BO: '591',
+  BL: '590',
+  BN: '673',
+  CC: '61',
+  CD: '243',
+  CI: '225',
+  KH: '855',
+  CM: '237',
+  // CA: '1',
+  CV: '238',
+  KY: '345',
+  CF: '236',
+  CH: '41',
+  CL: '56',
+  CN: '86',
+  CX: '61',
+  CO: '57',
+  KM: '269',
+  CG: '242',
+  CK: '682',
+  CR: '506',
+  CU: '53',
+  CY: '537',
+  CZ: '420',
+  DE: '49',
+  DK: '45',
+  DJ: '253',
+  DM: '1767',
+  DO: '1849',
+  DZ: '213',
+  EC: '593',
+  EG: '20',
+  ER: '291',
+  EE: '372',
+  ES: '34',
+  ET: '251',
+  FM: '691',
+  FK: '500',
+  FO: '298',
+  FJ: '679',
+  FI: '358',
+  FR: '33',
+  GB: '44',
+  GF: '594',
+  GA: '241',
+  GS: '500',
+  GM: '220',
+  GE: '995',
+  GH: '233',
+  GI: '350',
+  GQ: '240',
+  GR: '30',
+  GG: '44',
+  GL: '299',
+  GD: '1473',
+  GP: '590',
+  GU: '1671',
+  GT: '502',
+  GN: '224',
+  GW: '245',
+  GY: '595',
+  HT: '509',
+  HR: '385',
+  HN: '504',
+  HU: '36',
+  HK: '852',
+  IR: '98',
+  IM: '44',
+  IL: '972',
+  IO: '246',
+  IS: '354',
+  IN: '91',
+  ID: '62',
+  IQ: '964',
+  IE: '353',
+  IT: '39',
+  JM: '1876',
+  JP: '81',
+  JO: '962',
+  JE: '44',
+  KP: '850',
+  KR: '82',
+  KZ: '77',
+  KE: '254',
+  KI: '686',
+  KW: '965',
+  KG: '996',
+  KN: '1869',
+  LC: '1758',
+  LV: '371',
+  LB: '961',
+  LK: '94',
+  LS: '266',
+  LR: '231',
+  LI: '423',
+  LT: '370',
+  LU: '352',
+  LA: '856',
+  LY: '218',
+  MO: '853',
+  MK: '389',
+  MG: '261',
+  MW: '265',
+  MY: '60',
+  MV: '960',
+  ML: '223',
+  MT: '356',
+  MH: '692',
+  MQ: '596',
+  MR: '222',
+  MU: '230',
+  MX: '52',
+  MC: '377',
+  MN: '976',
+  ME: '382',
+  MP: '1670',
+  MS: '1664',
+  MA: '212',
+  MM: '95',
+  MF: '590',
+  MD: '373',
+  MZ: '258',
+  NA: '264',
+  NR: '674',
+  NP: '977',
+  NL: '31',
+  NC: '687',
+  NZ: '64',
+  NI: '505',
+  NE: '227',
+  NG: '234',
+  NU: '683',
+  NF: '672',
+  NO: '47',
+  OM: '968',
+  PK: '92',
+  PM: '508',
+  PW: '680',
+  PF: '689',
+  PA: '507',
+  PG: '675',
+  PY: '595',
+  PE: '51',
+  PH: '63',
+  PL: '48',
+  PN: '872',
+  PT: '351',
+  PR: '1939',
+  PS: '970',
+  QA: '974',
+  RO: '40',
+  RE: '262',
+  RS: '381',
+  RU: '7',
+  RW: '250',
+  SM: '378',
+  SA: '966',
+  SN: '221',
+  SC: '248',
+  SL: '232',
+  SG: '65',
+  SK: '421',
+  SI: '386',
+  SB: '677',
+  SH: '290',
+  SD: '249',
+  SR: '597',
+  SZ: '268',
+  SE: '46',
+  SV: '503',
+  ST: '239',
+  SO: '252',
+  SJ: '47',
+  SY: '963',
+  TW: '886',
+  TZ: '255',
+  TL: '670',
+  TD: '235',
+  TJ: '992',
+  TH: '66',
+  TG: '228',
+  TK: '690',
+  TO: '676',
+  TT: '1868',
+  TN: '216',
+  TR: '90',
+  TM: '993',
+  TC: '1649',
+  TV: '688',
+  UG: '256',
+  UA: '380',
+  US: '1',
+  UY: '598',
+  UZ: '998',
+  VA: '379',
+  VE: '58',
+  VN: '84',
+  VG: '1284',
+  VI: '1340',
+  VC: '1784',
+  VU: '678',
+  WS: '685',
+  WF: '681',
+  YE: '967',
+  YT: '262',
+  ZA: '27',
+  ZM: '260',
+  ZW: '263',
+  AQ: '672',
+};
