@@ -1,1 +1,1 @@
-declare module "react-native-keyboard-aware-scrollview"
+declare module 'react-native-keyboard-aware-scrollview';
