@@ -1,58 +1,5 @@
 export const Routes = {
   Auth: {
-    Login: 'Login',
-    SignUp: 'SignUp',
-    Otp: 'OtpScreen',
-    ForgetPass: 'ForgetPass',
-  },
-  GetStarting: {
-    onBoarding: 'onBoarding',
-  },
-  Home: {
-    homeScreen: 'home',
-  },
-  Detail: {
-    DishDetail: 'DishDetail',
-    ResDetail: 'ResDetails',
-  },
-
-  Orders: {
-    order: 'Orders',
-    orderStatus: 'OrderStatus',
-    OrderDetail: 'OrderDetail',
-  },
-  Offers: {
-    offer: 'OffersScreen',
-  },
-  More: {
-    more: 'More',
-    ProfileDetail: 'ProfileDetail',
-    EditProfile: 'EditProfile',
-    ContactUs: 'ContactUs',
-    Form: 'Form',
-  },
-  Container: {
-    main: 'Container',
-  },
-  ResturantList: {
-    resturants: 'ResturantsScreen',
-  },
-  Address: {
-    AddAddress: 'AddAddress',
-    SelectAddress: 'SelectAddress',
-  },
-  Cart: {
-    cartScreen: 'CartScreen',
-  },
-  CheckOut: {
-    checkOut: 'CheckOut',
-  },
-  Favourite: {
-    favList: 'Favourite',
-  },
-  Grocery: {
-    grocery: 'GroceryScreen',
-    category: 'category',
-    product: 'product',
+    login: 'login',
   },
 };
